requirements.txt
pbr
slack_sdk>=3.0
#slackeventsapi>=2.1.0
click
tabulate
python-dotenv
slack_bolt
aprsd>=3.0.0
oslo.config
rich
